Push all branches on deploy.
While this may push wip branches, the intent of deploy is to expose all git
references so that any may be requested by the qwerty.sh server.

Specifically, this is needed to checkout the release tag if the branch has
not been pushed, where `--all` is preferable to naming a branch (in case
the working branch is something other than 'master').