Support an arbitrary git revision with --ref.
The revision must be reachable in order to support a shallow clone. Use
the qwerty.sh --ref flag to indicate that the revision is an arbitrary
git revision (e.g. a short ref to a SHA such as ac0ffee) and therefore
clone the entire repository in order to support an exact reference.